Removed dependency on phyloseq
Previously a number of function examples depended on phyloseq since the CSS dataset was saved as a phyloseq object. I recently updated the css project and changed it so CSS is now a list.
As a result of removing this dependency, philr now passes check of package without error (still some warnings though).
+ added function named_rtree (not currently exported)

Signed-off-by: Justin Silverman <jsilve24@gmail.com>